Switch directory format
